+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-development:
+0.4.0:
   * add SQL logger to development environment
   * generate public dir when running generator
   * use tux as Sinatra console